Add ExprTk_BUILD_EXAMPLES to build examples
Also move examples into an examples folder
