+++ examples/exprtk_simple_example_03.cpp
@@ -2,7 +2,7 @@
  **************************************************************
  *         C++ Mathematical Expression Toolkit Library        *
  *                                                            *
- * Simple Example 17                                          *
+ * Simple Example 3                                           *
  * Author: Arash Partow (1999-2023)                           *
  * URL: https://www.partow.net/programming/exprtk/index.html  *
  *                                                            *
@@ -17,62 +17,43 @@
 
 
 #include <cstdio>
-#include <cstdlib>
-#include <ctime>
 #include <string>
 
 #include "exprtk.hpp"
 
-
 template <typename T>
-struct rnd_01 : public exprtk::ifunction<T>
-{
-   using exprtk::ifunction<T>::operator();
-
-   rnd_01() : exprtk::ifunction<T>(0)
-   { ::srand(static_cast<unsigned int>(time(NULL))); }
-
-   inline T operator()()
-   {
-      // Note: Do not use this in production
-      // Result is in the interval [0,1)
-      return T(::rand() / T(RAND_MAX + 1.0));
-   }
-};
-
-template <typename T>
-void monte_carlo_pi()
+void polynomial()
 {
    typedef exprtk::symbol_table<T> symbol_table_t;
    typedef exprtk::expression<T>   expression_t;
    typedef exprtk::parser<T>       parser_t;
 
-   const std::string monte_carlo_pi_program =
-                  " var experiments[5 * 10^7] := [(rnd_01^2 + rnd_01^2) <= 1]; "
-                  " 4 * sum(experiments) / experiments[];                      ";
+   const std::string expression_string =
+                  "25x^5 - 35x^4 - 15x^3 + 40x^2 - 15x + 1";
 
-   rnd_01<T> rnd01;
+   const T r0 = T(0);
+   const T r1 = T(1);
+         T  x = T(0);
 
    symbol_table_t symbol_table;
-   symbol_table.add_function("rnd_01",rnd01);
+   symbol_table.add_variable("x",x);
 
    expression_t expression;
    expression.register_symbol_table(symbol_table);
 
    parser_t parser;
-   parser.compile(monte_carlo_pi_program,expression);
+   parser.compile(expression_string,expression);
 
-   const T approximate_pi = expression.value();
+   const T delta = T(1.0 / 100.0);
 
-   const T real_pi = T(3.141592653589793238462643383279502); // or close enough...
-
-   printf("pi ~ %20.17f\terror: %20.17f\n",
-          approximate_pi,
-          std::abs(real_pi - approximate_pi));
+   for (x = r0; x <= r1; x += delta)
+   {
+      printf("%19.15f\t%19.15f\n", x, expression.value());
+   }
 }
 
 int main()
 {
-   monte_carlo_pi<double>();
+   polynomial<double>();
    return 0;
 }
