+++ examples/exprtk_simple_example_09.cpp
@@ -21,7 +21,6 @@
 
 #include "exprtk.hpp"
 
-
 template <typename T>
 void primes()
 {
