+++ examples/exprtk_simple_example_10.cpp
@@ -22,7 +22,6 @@
 
 #include "exprtk.hpp"
 
-
 template <typename T>
 void newton_sqrt()
 {
