+++ examples/exprtk_simple_example_13.cpp
@@ -23,7 +23,6 @@
 
 #include "exprtk.hpp"
 
-
 template <typename T>
 void savitzky_golay_filter()
 {
