+++ examples/exprtk_simple_example_15.cpp
@@ -21,7 +21,6 @@
 
 #include "exprtk.hpp"
 
-
 template <typename T>
 void black_scholes_merton_model()
 {
