+++ examples/exprtk_simple_example_17.cpp
@@ -2,7 +2,7 @@
  **************************************************************
  *         C++ Mathematical Expression Toolkit Library        *
  *                                                            *
- * Simple Example 5                                           *
+ * Simple Example 17                                          *
  * Author: Arash Partow (1999-2023)                           *
  * URL: https://www.partow.net/programming/exprtk/index.html  *
  *                                                            *
@@ -17,65 +17,61 @@
 
 
 #include <cstdio>
+#include <cstdlib>
+#include <ctime>
 #include <string>
 
 #include "exprtk.hpp"
 
-
 template <typename T>
-struct myfunc : public exprtk::ifunction<T>
+struct rnd_01 : public exprtk::ifunction<T>
 {
    using exprtk::ifunction<T>::operator();
 
-   myfunc()
-   : exprtk::ifunction<T>(2)
-   { exprtk::disable_has_side_effects(*this); }
+   rnd_01() : exprtk::ifunction<T>(0)
+   { ::srand(static_cast<unsigned int>(time(NULL))); }
 
-   inline T operator()(const T& v1, const T& v2)
+   inline T operator()()
    {
-      return T(1) + (v1 * v2) / T(3);
+      // Note: Do not use this in production
+      // Result is in the interval [0,1)
+      return T(::rand() / T(RAND_MAX + 1.0));
    }
 };
 
 template <typename T>
-inline T myotherfunc(T v0, T v1, T v2)
-{
-   return std::abs(v0 - v1) * v2;
-}
-
-template <typename T>
-void custom_function()
+void monte_carlo_pi()
 {
    typedef exprtk::symbol_table<T> symbol_table_t;
    typedef exprtk::expression<T>   expression_t;
    typedef exprtk::parser<T>       parser_t;
 
-   const std::string expression_string =
-                  "myfunc(sin(x / pi), otherfunc(3 * y, x / 2, x * y))";
+   const std::string monte_carlo_pi_program =
+                  " var experiments[5 * 10^7] := [(rnd_01^2 + rnd_01^2) <= 1]; "
+                  " 4 * sum(experiments) / experiments[];                      ";
 
-   T x = T(1);
-   T y = T(2);
-   myfunc<T> mf;
+   rnd_01<T> rnd01;
 
    symbol_table_t symbol_table;
-   symbol_table.add_variable("x",x);
-   symbol_table.add_variable("y",y);
-   symbol_table.add_function("myfunc",mf);
-   symbol_table.add_function("otherfunc",myotherfunc);
-   symbol_table.add_constants();
+   symbol_table.add_function("rnd_01",rnd01);
 
    expression_t expression;
    expression.register_symbol_table(symbol_table);
 
    parser_t parser;
-   parser.compile(expression_string,expression);
+   parser.compile(monte_carlo_pi_program,expression);
 
-   const T result = expression.value();
-   printf("Result: %10.5f\n",result);
+   const T approximate_pi = expression.value();
+
+   const T real_pi = T(3.141592653589793238462643383279502); // or close enough...
+
+   printf("pi ~ %20.17f\terror: %20.17f\n",
+          approximate_pi,
+          std::abs(real_pi - approximate_pi));
 }
 
 int main()
 {
-   custom_function<double>();
+   monte_carlo_pi<double>();
    return 0;
 }
